% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/its_llm.R
\name{its_llm}
\alias{its_llm}
\title{Estimates via Local Linear Regression with user-selected bandwidth}
\usage{
its_llm(df, rvar, outcome, trend = "none", bwL = Inf, bwR = Inf,
  donut = 0)
}
\arguments{
\item{df}{(required) \code{data.frame} containing all variables}

\item{rvar}{(required) the name of the running variable in \code{df}}

\item{outcome}{(required) the name of the outcome variable in \code{df}}

\item{trend}{include a linear term ('lin'), a quadratic term ('quad') or no trend at all ('none')?}

\item{bwL}{bandwidth selection to the left of the cut-point on the scale of \code{rvar}}

\item{bwR}{bandwidth selection to the right of the cut-point on the scale of \code{rvar}}

\item{donut}{length of the period after the cut-point for which the data are dropped (on the scale of \code{rvar}).}
}
\value{
\code{data.frame} with a single row and entries for the point estimate (\code{est}), 95\% confidence interval (\code{lo,hi}), 
standard error (\code{se}), p-value (\code{pval}) and the number of data points to the left/right of the cut-point used in the 
 estimation (\code{Nleft,Nright}).
}
\description{
\code{its_llm} estimates the intercept shift of a time series at a cut-point.
}
\details{
Estimates the size of the intercept shift of a time series at at cut-point (at zero) using 
a linear regression model with separate trends for the running variable to both sides
of the cut-point and within the neighborhood as defined by the bandwidth parameters (\code{bwL},\code{bwR}).
Standard errors are calculated based on the heteroskedasticity-consistent covariance matrix (HC3) from the \link{sandwich} package.

Use \code{\link{its_plot_samples}} to understand which data points are included when choosing different values for \code{bwL}, 
\code{bwR} and \code{donut}. 

When no values for \code{trend}, \code{bwL}, \code{bwR} and \code{donut} are supplied, the functions defaults to estimating the 
difference in means pooling all available data to the left/right of the cut-point.
}
\examples{
 \dontrun{
	
  	eventmonth <- -12:12
  	treat <- as.numeric(eventmonth >= 0)
  	y <- 1 + (1 * eventmonth) + treat*2 + rnorm(length(eventmonth))
  	df <- data.frame(y=y, eventmonth=eventmonth)
  	its_llm(df, rvar="eventmonth", outcome="y", trend='lin', bwL=4,bwR=4, donut=0) 		

		}

}
\seealso{
\code{\link{its_plot_samples}}, \code{\link{its_llm_placebo}}.
}

