% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itstools-package.R
\docType{package}
\name{itstools-package}
\alias{itstools}
\alias{itstools-package}
\title{ITS Tool Box}
\description{
Estimates via Local Linear Regression
}
\details{
Feedback is very welcome!


\tabular{ll}{
	Package: \tab itstools \cr
	Type: \tab Package\cr
	Version: \tab 0.0.1.0\cr
	Date: \tab 2017-02-20\cr
	License: \tab  GPL-3\cr
	}
}
\author{
Moritz Marbach \email{marbach@ipz.uzh.ch}
}
\references{

}

