% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/its_llm.R
\name{its_llm}
\alias{its_llm}
\title{ITS Estimates via Local Linear Regression}
\usage{
its_llm(df, rvar, outcome, trend = "none", bwL = Inf, bwR = Inf,
  donut = 0)
}
\arguments{
\item{df}{(required) \code{data.frame} containing all variables}

\item{rvar}{(required) a running variable in \code{df}}

\item{outcome}{(required) the outcome variable in \code{df}}

\item{trend}{can be 'lin' or 'quad' (linear or quadratic)}

\item{donut}{length of the period in after the cutpoint that is not used for the estimation}

\item{bw}{bandwidth selection on the scale of the \code{rvar}}
}
\description{
\code{its_llm} Estimates change in time series via local linear regression
}
\details{
tbd.
}
\examples{
 \dontrun{
	
  	eventmonth <- -12:12
  	treat <- as.numeric(eventmonth >= 0)
  	y <- 1 + (1 * eventmonth) + treat*2 + rnorm(length(eventmonth))
  	df <- data.frame(y=y, eventmonth=eventmonth)
  	its_llm(df, rvar="eventmonth", outcome="y", bwL=2,bwR=2, donut=0) 		

		}

}
\references{

}

