% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/its_llm.R
\name{its_llm}
\alias{its_llm}
\title{Estimates via Local Linear Regression with user-selected bandwidth}
\usage{
its_llm(df, rvar, outcome, trend = "none", bw = Inf, donut = 0,
  verbose = 0)
}
\arguments{
\item{df}{(required) \code{data.frame} containing all variables}

\item{rvar}{(required) the name of the running variable in \code{df}}

\item{outcome}{(required) the name of the outcome variable in \code{df}}

\item{trend}{include a linear term ('lin'), a quadratic term ('quad') or no trend at all ('none')?}

\item{bw}{either a scalar or a vector of length 2 defining the bandwidth to the left (right) of the cut-point on the scale of \code{rvar}}

\item{donut}{either a scalar or a vector of length 2 defining the length of the period to the left (right) of the cut-point for which the data are dropped (on the scale of \code{rvar}).}

\item{verbose}{set to any value other than zero to show which data points are included in the estimation}
}
\value{
\code{data.frame} with a single row and entries for the point estimate (\code{est}), 95\% confidence interval (\code{lo,hi}), 
standard error (\code{se}), p-value (\code{pval}) and the number of data points to the left/right of the cut-point used in the 
estimation (\code{Nleft,Nright}).
}
\description{
\code{its_llm} estimates the intercept shift of a time series at a cut-point.
}
\details{
Estimates the size of the intercept shift of a time series at at cut-point (at zero) using 
a linear regression model with separate trends for the running variable to both sides
of the cut-point and within the neighborhood as defined by the bandwidth parameters (\code{bwL},\code{bwR}).
Standard errors are calculated based on the heteroskedasticity-consistent covariance matrix (HC3) from the \link{sandwich} package.

Use \code{\link{its_plot_samples}} to understand which data points are included when choosing different values for \code{bwL}, 
\code{bwR} and \code{donut}. 

When no values for \code{trend}, \code{bwL}, \code{bwR} and \code{donut} are supplied, the functions defaults to estimating the 
difference in means pooling all available data to the left/right of the cut-point.
}
\examples{
\dontrun{
	
  N <- 21
  time <- seq(-1,1,length.out=N)
  treat <- as.numeric(time >= 0)
  y <- 1 + time + treat*1 + rnorm(N,0,0.25)
  df <- data.frame(y=y, time=time)
  its_llm(df, rvar="time", outcome="y", bw=0.25)

}

}
\seealso{
\code{\link{its_plot_samples}}, \code{\link{its_llm_placebo}}.
}

