% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/its_llm_placebo.R
\name{its_llm_placebo}
\alias{its_llm_placebo}
\title{ITS Placebo Estimates}
\usage{
its_llm_placebo(df, rvar, outcome, trend = NULL, bw, donut = 0,
  nsim = 200)
}
\arguments{
\item{df}{(required) \code{data.frame} containing all variables}

\item{rvar}{(required) a running variable in \code{df}}

\item{outcome}{(required) the outcome variable in \code{df}}

\item{trend}{can be 'lin' or 'quad' (linear or quadratic)}

\item{bw}{bandwidth selection on the scale of the \code{rvar}}

\item{donut}{length of the period in after the cutpoint that is not used for the estimation}

\item{nsim}{number of simulations to run}
}
\description{
\code{its_llm_placebo} gives a vector of placebo estimates
}
\details{
tbd.
}
\examples{
 \dontrun{

}

}
\references{

}

