% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/its_llm_plot.R
\name{its_llm_plot}
\alias{its_llm_plot}
\title{Plot the Estimation and Placebo Sample}
\usage{
its_llm_plot(df, rvar, outcome, bwL, bwR, donut)
}
\arguments{
\item{df}{(required) \code{data.frame} containing all variables}

\item{rvar}{(required) the name of the running variable in \code{df}}

\item{outcome}{(required) the name of the outcome variable in \code{df}}

\item{bwL}{bandwidth selection to the left of the cut-point on the scale of \code{rvar}}

\item{bwR}{bandwidth selection to the right of the cut-point on the scale of \code{rvar}}

\item{donut}{length of the period after the cut-point for which the data are dropped (on the scale of \code{rvar}).}

\item{trend}{include a linear term ('lin'), a quadratic term ('quad') or no trend at all ('none')?}
}
\description{
\code{its_llm_plot} helps to visualize which data points are used to estimate the intercept shift 
at the cut-point and the placebo estimates.
}
\examples{
 \dontrun{

  eventmonth <- -12:12
  treat <- as.numeric(eventmonth >= 0)
  y <- 1 + (1 * eventmonth) + treat*2 + rnorm(length(eventmonth))
  df <- data.frame(y=y, eventmonth=eventmonth)
  its_llm_plot(df, rvar="eventmonth", outcome="y", bwL=2,bwR=2, donut=0)

}

}
\references{

}
\seealso{
\code{\link{its_llm_placebo}}, \code{\link{its_llm}}.
}

