% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/its_loess.R
\name{its_loess}
\alias{its_loess}
\title{ITS Estimates via Local Polynomial Regression}
\usage{
its_loess(df, rvar, outcome, span = 0.8, degree = 1, othervar = NULL,
  only_estimates = TRUE)
}
\arguments{
\item{df}{(required) \code{data.frame} containing all variables}

\item{rvar}{(required) a running variable in \code{df}}

\item{outcome}{(required) the outcome variable in \code{df}}

\item{trend}{can be 'lin' or 'quad' (linear or quadratic)}

\item{bw}{bandwidth selection on the scale of the \code{rvar}}

\item{donut}{length of the period in after the cutpoint that is not used for the estimation}
}
\description{
\code{its_llm} Estimates change in time series via local linear regression
}
\details{

}
\examples{
 \dontrun{
	
  	eventmonth <- -12:12
  	treat <- as.numeric(eventmonth >= 0)
  	y <- 1 + (1 * eventmonth) + treat*2 + rnorm(length(eventmonth))
  	df <- data.frame(y=y, eventmonth=eventmonth)
  	its_loess(df, rvar="eventmonth", outcome="y") 		

		}

}
\references{

}

