% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/its_plot_samples.R
\name{its_plot_samples}
\alias{its_plot_samples}
\title{Plots ITS Samples}
\usage{
its_plot_samples(df, rvar, outcome, bw, donut)
}
\arguments{
\item{df}{(required) \code{data.frame} containing all variables}

\item{rvar}{(required) a running variable in \code{df}}

\item{outcome}{(required) the outcome variable in \code{df}}

\item{bw}{bandwidth selection on the scale of the \code{rvar}}

\item{donut}{length of the period in after the cutpoint that is not used for the estimation}
}
\description{
\code{its_plot_samples} ...
}
\details{
tbd.
}
\examples{
 \dontrun{

  eventmonth <- sort(c(0,runif(1000,-1,1)))
  treat <- as.numeric(eventmonth >= 0)
  y <- 1 + (1 * eventmonth) + treat*2 + rnorm(length(eventmonth))
  df <- data.frame(y=y, eventmonth=eventmonth)
  its_plot_samples(df, rvar="eventmonth", outcome="y", bw=0.25, donut=0)

}

}
\references{

}

